+++ thelper/cli.py
@@ -9,7 +9,6 @@ outputs, the path to a directory where to save the data is also needed.
 """
 
 import argparse
-import json
 import logging
 import os
 from pathlib import Path as pth
@@ -346,7 +345,7 @@ def inference_session(config, save_dir=None, ckpt_path=None):
         raise RuntimeError("Model checkpoint was explicitly defined as not pretrained. It must be for inference.")
 
     if not os.path.exists(ckpt_path):
-        logger.fatal(f"Model not found: {ckpt_path}")
+        logger.fatal("Model not found: %s", ckpt_path)
         raise AssertionError("Model checkpoint missing to run inference")
     ckptdata = thelper.utils.load_checkpoint(ckpt_path, map_location=None, always_load_latest=False)
     if "task" not in ckptdata or not isinstance(ckptdata["task"], (thelper.tasks.Task, str)):
@@ -402,17 +401,17 @@ def inference_session(config, save_dir=None, ckpt_path=None):
 
     config_file = "config-train.json"
     config_file_path = os.path.join(save_dir, config_file)
-    train_config = ckptdata['config']
+    train_config = ckptdata["config"]
     logger.debug("Writing employed train config: [%s]", config_file_path)
-    with open(config_file_path, 'w') as f:
-        json.dump(train_config, f, indent=4)
+    thelper.utils.save_config(train_config, config_file_path, force_convert=True)
     config_name = "config-infer.json"
     config_file_path = os.path.join(save_dir, config_name)
     logger.debug("Writing employed infer config: [%s]", config_file_path)
-    with open(config_file_path, 'w') as f:
-        json.dump(config, f, indent=4)
+    thelper.utils.save_config(config, config_file_path, force_convert=True)
 
+    logger.info("Starting inference.")
     tester.test()
+    logger.info("Inference completed.")
 
 
 def export_model(config, save_dir):