+++ thelper/data/geo/parsers.py
@@ -8,10 +8,7 @@ import os
 import pickle
 
 import cv2 as cv
-import gdal
 import numpy as np
-import ogr
-import osr
 import shapely
 import torch
 import tqdm
@@ -20,6 +17,13 @@ import thelper.tasks
 import thelper.utils
 from thelper.data import Dataset, ImageFolderDataset
 from thelper.data.geo.utils import parse_raster_metadata
+
+try:
+    from osgeo import gdal, ogr, osr
+except (ImportError, ModuleNotFoundError):
+    import gdal
+    import ogr
+    import osr
 
 logger = logging.getLogger(__name__)